requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.3
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
whitenoise==6.2.0
